Added a new component and a new animation for settings
