+++ .env
@@ -10,3 +10,4 @@ BREVO_USER=user
 BREVO_PASSWORD=pass
 SUPPORT_FROM=AICulture-Support@myapp.com
 OPENWEATHER_API=681a847ebddfec8c90bc96ae7e0af34e
+OPENAI_API_KEY=key
